AVLTree Song
This project, gets the songs from a txt file and adds them to the an AVL Tree. Then we can search the songs in 2 type. 1st type has tree options that are by name, by artist and by ID of the song. 2nd type is searching between 2 ID(upper and lower bound).